Initial multi statement match prototype
Type checks, and is able to connect two statement matches based on a
high level EDSL which would need more work. This should be generalized
to allow more than two points, possibly other combinations.

Graph is explored from a match point going forward with a cap that can
be same block, function or some other means. It would need better such
primitives as one may or may not want to explore into called functions
which is not entirely in place yet, but the general idea is.

I also think one wants to have some kind of summary graphs and indexing
for performance. To allow quickly finding locations far away and to
quickly update based on summary in between.